refactor: DateFormat moved out of jersey package
